Updated project to clear up all compiler warnings when building.
- For all 3rd party files which we prefixed ourselves, I have just ignored all the warnings as changing 3rd party code is never recommended.

- For all of our own codes warnings,  I have gone through and made the necessary changes to fix the warnings.

+++ source/Classes/UIEventTracking/UIApplication+ApigeeSwizzling.m
@@ -70,7 +70,7 @@ static void OverrideUIApplicationSendAction(UIApplication* self, SEL _cmd, SEL a
             
             if( barButtonItem.tag > 0 )
             {
-                buttonPressEvent.tag = barButtonItem.tag;
+                buttonPressEvent.tag = (int) barButtonItem.tag;
             }
             
             [[ApigeeUIEventManager sharedInstance] notifyButtonPress:buttonPressEvent];
